Refactor HTML attributes and node class structures
In this commit, the HTML attribute and node classes have been significantly refactored. The attribute interface 'IAttribute' was removed, 'IBooleanAttribute' was renamed to 'BooleanAttribute', 'IGlobalAttribute' was current to 'GlobalAttribute' and 'IRequiresAttribute' was changed to 'RequiresAttribute'. Additionally, 'Node' was migrated to 'Element' and related class imports and variables were updated accordingly. Changes were applied throughout the code including different attribute classes and element implementations.